refactor: move ECC_Context to ecc_context.{h,cpp}
No behavior change.

ECC_Context now lives in the consensus library, so it can
hold the libsecp verification context next.
