+++ src/key.cpp
@@ -7,6 +7,7 @@
 
 #include <crypto/common.h>
 #include <crypto/hmac_sha512.h>
+#include <ecc_context.h>
 #include <hash.h>
 #include <random.h>
 
@@ -18,8 +19,6 @@
 
 #include <algorithm>
 #include <span>
-
-static const ECC_Context* g_ecc_context = nullptr;
 
 /** These functions are taken from the libsecp256k1 distribution and are very ugly. */
 
@@ -452,39 +451,4 @@ bool KeyPair::SignSchnorr(const uint256& hash, std::span<unsigned char> sig, con
     }
     if (!ret) memory_cleanse(sig.data(), sig.size());
     return ret;
-}
-
-secp256k1_context* GetSecp256k1SignContext()
-{
-    return g_ecc_context ? g_ecc_context->SignContext() : nullptr;
-}
-
-/** Create an elliptic curve context. Provide rng seed for blinding factor if needed */
-static secp256k1_context* ECC_Start(std::span<const unsigned char> rng_seed32) {
-    assert(rng_seed32.empty() || rng_seed32.size() == 32);
-
-    secp256k1_context *ctx = secp256k1_context_create(SECP256K1_CONTEXT_NONE);
-    assert(ctx != nullptr);
-
-    if (!rng_seed32.empty()){
-        // Pass in a random blinding seed to the secp256k1 context.
-        bool ret = secp256k1_context_randomize(ctx, rng_seed32.data());
-        assert(ret);
-    }
-
-    return ctx;
-}
-
-ECC_Context::ECC_Context(std::span<const unsigned char> rng_seed32)
-{
-    assert(g_ecc_context == nullptr);
-    m_sign_ctx = ECC_Start(rng_seed32);
-    g_ecc_context = this;
-}
-
-ECC_Context::~ECC_Context()
-{
-    assert(g_ecc_context == this);
-    g_ecc_context = nullptr;
-    secp256k1_context_destroy(m_sign_ctx);
 }
